+++ src/App.tsx
@@ -17,6 +17,8 @@ import NewListDialog from "./components/NewListDialog";
 import HelpDialog from "./components/HelpDialog";
 import ConfirmDialog from "./components/ConfirmDialog";
 import ScanProgress from "./components/ScanProgress";
+import SelectionBar from "./components/SelectionBar";
+import RowMenu from "./components/RowMenu";
 import SheetDialog from "./components/SheetDialog";
 import ServiceView from "./components/ServiceView";
 import Toast from "./components/Toast";
@@ -119,6 +121,8 @@ export default function App() {
       <ConfirmDialog />
       <ServiceView />
       <SheetDialog />
+      <SelectionBar />
+      <RowMenu />
       <ScanProgress />
       <Toast />
     </div>